test(sip): move OPTIONS test out of the TLS file and poll for readiness
    Follow-up cleanup to #94. Two test-hygiene fixes, no behaviour change:

    TestEngine_OPTIONSReturnsOK landed in engine_tls_test.go but exercises
    nothing TLS-related, so it moves next to TestEngine_AllowHeader — the
    assertion it complements. pickFreePort is package-scoped, so the move is
    free; engine_tls_test.go drops the sipgo imports it no longer needs.

    The test also waited for the listener with a fixed 100ms sleep, which is
    a flake on a loaded CI box. It now retries real OPTIONS requests against
    a 3s deadline, matching the poll-until-ready loop already used by
    TestEngine_Serve_AcceptsTLSHandshake. This is not theoretical: under
    -count=5 one run took 550ms to get its first answer, which the old sleep
    would have failed. Each attempt builds a fresh request because sipgo
    populates Via and Call-ID in place.
